examples/google-adk/requirements.txt
google-adk>=0.1.0
sidclaw>=0.1.2
python-dotenv>=1.0.0
